nbdkit: Add stats filter in verbose mode
If we run in verbose mode, adds the stats filter before all other
plugins. This shows what qemu-img convert is seeing both on input and
output side, and may help to analyze performance issues.

Starting with the simplest thing that can work, writing the stats to
stderr, where nbdkit logs.