output/rhv-upload-plugin: Keep connections alive
When importing from vddk, nbdcopy may be blocked for few minutes(!)
trying to get extents. While nbdcopy is blocked, imageio server closes
the idle connections. When we finally get a request from nbdcopy, we
fail to detect that the connection was closed.

Detecting a closed connection is hard and racy. In the good case, we get
a BrokenPipe error. In the bad case, imageio closed the socket right
after we sent a request, and we get an invalid status line.

Even worse, when we find that the connection was closed, it is not safe
to reopen the connection, since nobody ensure that data written to the
previous connection will be flushed when we flush the new connection.

Fix the issue by keeping the connections alive. A pool keeper thread
sends a flush request on idle connection every 30 seconds.

When we importing from fast input (e.g. local file), the keeper thread
will never find an idle connection. When importing from slow input
(vddk), we will send a flush request on every connection when nbdcopy is
busy waiting for vddk.

Fixes https://bugzilla.redhat.com/2032324