output/rhv-upload-plugin: Fix flush and close
When locking the http pool, we wait until all connections are idle, and
take them from the pool. But since we used pool.qsize(), which is the
number of items currently in the queue, we did not wait for all
connections.

This leads to following issues:

- We send flush request only for some connections, which does not ensure
  that all uploaded data is flushed to storage.

- We close only some of the connections in cleanup(). This should not
  matter since the connections are closed when the plugin process
  terminates.

An example import showing sending only one FLUSH request instead of 4:
https://bugzilla.redhat.com/2032324#c8

Fixed by creating a bounded queue and using pool.maxsize to get all the
connections from the pool.